requirements.txt

# robotframework related
robotframework
robotframework-seleniumlibrary
robotframework-screencaplibrary
selenium
pyhocon
jinja2
requests
requests_toolbelt
python-crontab
croniter
paho-mqtt
websocket_client
boto
jsonpath
chromedriver-binary
toolz

pycryptodome
xlrd
xlwt
pymongo
influxdb
PyYAML
python-dateutil
elasticsearch

pykafka
kafka-python==1.4.7

blinker
pyjwt
cryptography
bs2json
beautifulsoup4
